chore: update openapi prefix
