+++ server/acl.go
@@ -35,8 +35,8 @@ func aclMiddleware(s *Server, ce *casbin.Enforcer, next echo.HandlerFunc, readon
 		if common.HasPrefixes(c.Path(), "/api/subscription") && method == "GET" {
 			return next(c)
 		}
-		// Skip GET /api/sql/advise request
-		if common.HasPrefixes(c.Path(), "/api/sql/advise") && method == "GET" {
+		// Skip OpenAPI request
+		if common.HasPrefixes(c.Path(), openAPIPrefix) {
 			return next(c)
 		}
